order-service/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
exceptiongroup==1.3.0
fastapi==0.115.12
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pydantic==2.11.5
pydantic_core==2.33.2
pytest==8.3.5
python-dotenv==1.1.0
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
